create a validate subcommand
